Updating document type service for slugify
